# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [identity]
      summary: Get signed-in user context
      responses:
        "200":
          description: Current user and organisation context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: List import runs
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Organisation-scoped import runs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
    post:
      operationId: createRun
      tags: [runs]
      summary: Create a queued import run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunInput"
      responses:
        "201":
          description: Queued import run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Invalid file metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /runs/{runId}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Get an import run
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Import run detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Run not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /runs/{runId}/process:
    post:
      operationId: processRun
      tags: [runs]
      summary: Process a queued import run
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: Run processing accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "409":
          description: Run cannot be processed in its current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /runs/{runId}/exceptions:
    get:
      operationId: listRunExceptions
      tags: [runs]
      summary: List validation exceptions for a run
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Validation exceptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationException"
  /runs/{runId}/summary:
    get:
      operationId: getRunSummary
      tags: [runs]
      summary: Get the structured evidence summary
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Summary detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceSummary"
        "404":
          description: Summary not generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: generateRunSummary
      tags: [runs]
      summary: Generate a structured summary using the LLM
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryInput"
      responses:
        "201":
          description: Generated evidence summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceSummary"
        "409":
          description: Run is not ready for summarisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /actions:
    get:
      operationId: listActions
      tags: [actions]
      summary: List follow-up actions
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [requested, approved, rejected, running, completed, failed]
      responses:
        "200":
          description: Organisation-scoped actions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActionRequest"
    post:
      operationId: requestAction
      tags: [actions]
      summary: Request a follow-up action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionRequestInput"
      responses:
        "201":
          description: Action awaiting administrator approval.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequest"
  /actions/{actionId}/approve:
    post:
      operationId: approveAction
      tags: [actions]
      summary: Approve and run a follow-up action
      parameters:
        - $ref: "#/components/parameters/ActionId"
      responses:
        "200":
          description: Approved action result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequest"
        "403":
          description: Administrator role required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /actions/{actionId}/reject:
    post:
      operationId: rejectAction
      tags: [actions]
      summary: Reject a follow-up action
      parameters:
        - $ref: "#/components/parameters/ActionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionDecisionInput"
      responses:
        "200":
          description: Rejected action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequest"
  /audit:
    get:
      operationId: listAuditEvents
      tags: [audit]
      summary: List the organisation audit trail
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Immutable audit events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get organisation settings
      responses:
        "200":
          description: Organisation settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update organisation settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated organisation settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationSettings"
        "403":
          description: Administrator role required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ActionId:
      name: actionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Role:
      type: string
      enum: [analyst, administrator, auditor]
    RunStatus:
      type: string
      enum: [queued, running, succeeded, failed, partial]
    CurrentUser:
      type: object
      required: [id, name, email, role, organisation]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        organisation:
          $ref: "#/components/schemas/Organisation"
...
    Run:
      type: object
      required: [id, fileName, fileType, status, recordCount, exceptionCount, retryCount, createdAt, summaryStatus]
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        fileType: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        recordCount: { type: integer }
        exceptionCount: { type: integer }
        retryCount: { type: integer }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        summaryStatus:
          type: string
          enum: [not_started, generating, ready, failed]
    RunInput:
      type: object
      required: [fileName, fileType, fileSize, objectPath]
      properties:
        fileName: { type: string, minLength: 1 }
        fileType: { type: string, minLength: 1 }
        fileSize: { type: integer, minimum: 1 }
        objectPath: { type: string, minLength: 1 }
        idempotencyKey: { type: string, minLength: 1 }
    ValidationException:
      type: object
      required: [id, runId, rowNumber, field, code, message, severity, status]
      properties:
        id: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        rowNumber: { type: integer }
        field: { type: string }
        code: { type: string }
        message: { type: string }
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, reviewed, waived]
        value:
          type: ["string", "null"]
    SummaryInput:
      type: object
      properties:
        forceRegenerate: { type: boolean, default: false }
    EvidenceSummary:
      type: object
      required: [id, runId, headline, overview, riskLevel, findings, generatedAt, model]
      properties:
        id: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        headline: { type: string }
        overview: { type: string }
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        findings:
          type: array
          items:
            type: object
            required: [title, detail, severity]
            properties:
              title: { type: string }
              detail: { type: string }
              severity:
                type: string
                enum: [low, medium, high, critical]
        generatedAt: { type: string, format: date-time }
        model: { type: string }
    ActionRequest:
      type: object
      required: [id, runId, actionType, title, rationale, status, requestedBy, requestedAt]
      properties:
        id: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        actionType:
          type: string
          enum: [request_correction, notify_owner, create_review_task]
        title: { type: string }
        rationale: { type: string }
        status:
          type: string
          enum: [requested, approved, rejected, running, completed, failed]
        requestedBy: { type: string }
        requestedAt: { type: string, format: date-time }
        decidedAt: { type: ["string", "null"], format: date-time }
        decisionNote: { type: ["string", "null"] }
        result: { type: ["string", "null"] }
    ActionRequestInput:
      type: object
      required: [runId, actionType, title, rationale]
      properties:
        runId: { type: string, format: uuid }
        actionType:
          type: string
          enum: [request_correction, notify_owner, create_review_task]
        title: { type: string, minLength: 1 }
        rationale: { type: string, minLength: 1 }
    ActionDecisionInput:
      type: object
      properties:
        note: { type: string }
    AuditEvent:
      type: object
      required: [id, action, entityType, entityId, actor, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string }
        actor: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        createdAt: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: true }
    OrganisationSettings:
      type: object
      required: [id, name, code, retentionDays, requireApproval]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        retentionDays: { type: integer, minimum: 1 }
        requireApproval: { type: boolean }
...